Refactor search
