+++ ts/packages/knowPro/test/testCommon.ts
@@ -20,6 +20,7 @@ import {
     IMessage,
     ITermToSemanticRefIndex,
     SemanticRef,
+    PropertySearchTerm,
 } from "../src/interfaces.js";
 import {
     ConversationSettings,
@@ -27,11 +28,7 @@ import {
 } from "../src/conversation.js";
 import { createConversationFromData } from "../src/common.js";
 import { readConversationDataFromFile } from "../src/serialization.js";
-import {
-    createSearchTerm,
-    PropertySearchTerm,
-    SemanticRefSearchResult,
-} from "../src/search.js";
+import { createSearchTerm, SemanticRefSearchResult } from "../src/search.js";
 import * as q from "../src/query.js";
 import { PropertyNames } from "../src/propertyIndex.js";
 import { Result } from "typechat";
